move plot painting logic
